Fixed name of Damageable
