+++ Assets/Scripts/Damageable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cdb3495ee6d6e4fbd84451ec7d431996
+guid: 4e0f89f6e5477473eb0174803c2ece05
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
